Implement particle limit beyond which emitters will stop emitting

+++ src/gizmos/emitter.rs
@@ -1,8 +1,10 @@
 use bevy::prelude::*;
+use leafwing_input_manager::prelude::ActionState;
 
 use crate::{
 	common::Positive,
-	particle::{spawn_particle_at_location, NextBatch},
+	input::Action,
+	particle::{spawn_particle_at_location, NextBatch, Particle},
 };
 
 #[derive(Component)]
@@ -29,9 +31,15 @@ impl Default for Emitter {
 pub fn activate_emitters(
 	mut commands: Commands,
 	time: Res<Time>,
+	limit: Res<ParticleLimit>,
 	mut next_batch: ResMut<NextBatch>,
+	particles: Query<(), With<Particle>>,
 	mut emitters: Query<(&mut Emitter, Option<&Positive>, &Transform)>,
 ) {
+	if particles.iter().len() >= limit.current() as usize {
+		return;
+	}
+
 	for (mut emitter, positive, transform) in &mut emitters {
 		let location = transform.translation.truncate();
 		if emitter.time_since_emitting > emitter.interval {
@@ -45,5 +53,41 @@ pub fn activate_emitters(
 		} else {
 			emitter.time_since_emitting += time.delta_seconds();
 		}
+	}
+}
+
+pub struct ParticleLimit(u32);
+
+impl ParticleLimit {
+	fn raise(&mut self) {
+		self.0 = self.0.saturating_add(100).min(u32::MAX - u32::MAX % 100);
+	}
+	fn lower(&mut self) {
+		self.0 = self.0.saturating_sub(100);
+	}
+	pub fn current(&self) -> u32 {
+		self.0
+	}
+}
+
+impl Default for ParticleLimit {
+	fn default() -> Self {
+		Self(1_200)
+	}
+}
+
+pub fn adjust_particle_limit(
+	mut limit: ResMut<ParticleLimit>,
+	action_state: Query<&ActionState<Action>>,
+) {
+	let action_state = action_state.single();
+
+	match (
+		action_state.just_pressed(Action::RaiseParticleLimit),
+		action_state.just_pressed(Action::LowerParticleLimit),
+	) {
+		(true, false) => limit.raise(),
+		(false, true) => limit.lower(),
+		_ => (),
 	}
 }
